refactor out heading component
