+++ client/components/BeerTransactionList.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import classNames from 'classnames'
 import BeerTransaction from './BeerTransaction'
+import Heading from './Heading'
 
 class BeerTransactionList extends React.Component {
   constructor(props) {
@@ -8,26 +9,16 @@ class BeerTransactionList extends React.Component {
   }
 
   render() {
-    const header = classNames('transaction')
-    const flexTitle = classNames('flex-title')
-    const container = classNames('flex-container')
-    const heading = classNames('inventory-headings', 'small-margins')
     return (
       <div>
-        <div className={container}>
-          <div className={header}>
-          <div className={flexTitle}>
-            <h2 className={heading}>Inventory Transactions</h2>
-          </div>
-        </div>
-      </div>
+        <Heading text="Inventory Transactions" />
         <ul>
           {
             this.props.transactions.length > 0 ?
             this.props.transactions.map((transaction) =>
               <BeerTransaction transaction={transaction} key={transaction.id}/>
             ) :
-            <h2 >No Transactions Found</h2>
+            <Heading text="No Transactions Found" />
           }
         </ul>
       </div>
